total amount calculate
expenses
                .reduce((acc, expense) => expense.amount + acc, 0)
                .toFixed(2)

reduce we use to combine all the elements in the array to something else.

inhere we have to use arrow function with 2 parameters,  acc is stand for the variable called accumilator it is just variable second one is expenses object in arrow function we combine the acc+expenses.amount together. we can define accumilator default value and it represented there in with 0.

.toFixed(2) is used for the total amount with 2 decimal values like 22.00 .

00 from the toFixed(2)